Update on "Add Parameter Server Example to CI Tests"
[ghstack-poisoned]
