+++ torch/testing/_internal/distributed/rpc/examples/parameter_server_test.py
@@ -1,139 +1,268 @@
 # If you need to modify this file to make this test pass, please also apply same edits accordingly to
-# https://github.com/pytorch/examples/blob/master/distributed/rpc/batch/parameter_server.py
-# and https://pytorch.org/tutorials/intermediate/rpc_async_execution.html#batch-updating-parameter-server
-
-import threading
-from datetime import datetime
-from time import perf_counter
+# https://github.com/pytorch/examples/blob/master/distributed/rpc/parameter_server/rpc_parameter_server.py
+# and https://pytorch.org/tutorials/intermediate/rpc_param_server_tutorial.html
 
 import torch
-import torch.distributed.rpc as rpc
 import torch.nn as nn
+import torch.nn.functional as F
 from torch import optim
+import torch.distributed.rpc as rpc
+import torch.distributed.autograd as dist_autograd
+from torch.distributed.optim import DistributedOptimizer
+from threading import Lock
 
-from torch.testing._internal.dist_utils import (
-    dist_init,
-    worker_name,
-)
+from torch.testing._internal.dist_utils import dist_init
 from torch.testing._internal.distributed.rpc.rpc_agent_test_fixture import RpcAgentTestFixture
 
-batch_size = 20
-in_features = 100
-out_features = 30
-num_batches = 4
+# --------- MNIST Network to train, from pytorch/examples -----
 
 
-def timed_log(text):
-    print(f"{datetime.now().strftime('%H:%M:%S')} {text}")
+class Net(nn.Module):
+    def __init__(self, num_gpus=0):
+        super(Net, self).__init__()
+        print(f"Using {num_gpus} GPUs to train")
+        self.num_gpus = num_gpus
+        device = torch.device(
+            "cuda:0" if torch.cuda.is_available() and self.num_gpus > 0 else "cpu")
+        print(f"Putting first 2 convs on {str(device)}")
+        # Put conv layers on the first cuda device
+        self.conv1 = nn.Conv2d(1, 32, 3, 1).to(device)
+        self.conv2 = nn.Conv2d(32, 64, 3, 1).to(device)
+        # Put rest of the network on the 2nd cuda device, if there is one
+        if "cuda" in str(device) and num_gpus > 1:
+            device = torch.device("cuda:1")
+
+        print(f"Putting rest of layers on {str(device)}")
+        self.dropout1 = nn.Dropout2d(0.25).to(device)
+        self.dropout2 = nn.Dropout2d(0.5).to(device)
+        self.fc1 = nn.Linear(9216, 128).to(device)
+        self.fc2 = nn.Linear(128, 10).to(device)
+
+    def forward(self, x):
+        x = self.conv1(x)
+        x = F.relu(x)
+        x = self.conv2(x)
+        x = F.max_pool2d(x, 2)
+
+        x = self.dropout1(x)
+        x = torch.flatten(x, 1)
+        # Move tensor to next device if necessary
+        next_device = next(self.fc1.parameters()).device
+        x = x.to(next_device)
+
+        x = self.fc1(x)
+        x = F.relu(x)
+        x = self.dropout2(x)
+        x = self.fc2(x)
+        output = F.log_softmax(x, dim=1)
+        return output
 
 
-class BatchUpdateParameterServer(object):
+# --------- Helper Methods --------------------
 
-    def __init__(self, batch_update_size):
-        self.model = nn.Linear(in_features, out_features)
-        self.lock = threading.Lock()
-        self.future_model = torch.futures.Future()
-        self.batch_update_size = batch_update_size
-        self.curr_update_size = 0
-        self.optimizer = optim.SGD(self.model.parameters(), lr=0.001, momentum=0.9)
-        for p in self.model.parameters():
-            p.grad = torch.zeros_like(p)
+# On the local node, call a method with first arg as the value held by the
+# RRef. Other args are passed in as arguments to the function called.
+# Useful for calling instance methods.
+def call_method(method, rref, *args, **kwargs):
+    return method(rref.local_value(), *args, **kwargs)
 
-    def get_model(self):
-        return self.model
-
-    @staticmethod
-    @rpc.functions.async_execution
-    def update_and_fetch_model(ps_rref, grads):
-        self = ps_rref.local_value()
-        for p, g in zip(self.model.parameters(), grads):
-            p.grad += g
-        with self.lock:
-            timed_log(f"PS got {self.curr_update_size}/{self.batch_update_size} updates")
-            self.curr_update_size += 1
-            fut = self.future_model
-
-            if self.curr_update_size >= self.batch_update_size:
-                for p in self.model.parameters():
-                    p.grad /= self.batch_update_size
-                self.curr_update_size = 0
-                self.optimizer.step()
-                self.optimizer.zero_grad()
-                fut.set_result(self.model)
-                timed_log("PS updated model")
-                self.future_model = torch.futures.Future()
-
-        return fut
+# Given an RRef, return the result of calling the passed in method on the value
+# held by the RRef. This call is done on the remote node that owns
+# the RRef. args and kwargs are passed into the method.
+# Example: If the value held by the RRef is of type Foo, then
+# remote_method(Foo.bar, rref, arg1, arg2) is equivalent to calling
+# <foo_instance>.bar(arg1, arg2) on the remote node and getting the result
+# back.
 
 
-class Trainer(object):
-
-    def __init__(self, ps_rref):
-        self.ps_rref = ps_rref
-        self.loss_fn = nn.L1Loss()
-
-    def get_next_batch(self):
-        for _ in range(num_batches):
-            inputs = torch.randn(batch_size, in_features)
-            labels = torch.zeros(batch_size, out_features)
-            yield inputs, labels
-
-    def train(self):
-        name = rpc.get_worker_info().name
-        m = self.ps_rref.rpc_sync().get_model()
-        for inputs, labels in self.get_next_batch():
-            timed_log(f"{name} processing one batch")
-            self.loss_fn(m(inputs), labels).backward()
-            timed_log(f"{name} reporting grads")
-            m = rpc.rpc_sync(
-                self.ps_rref.owner(),
-                BatchUpdateParameterServer.update_and_fetch_model,
-                args=(self.ps_rref, [p.grad for p in m.cpu().parameters()]),
-            )
-            timed_log(f"{name} got updated model")
+def remote_method(method, rref, *args, **kwargs):
+    args = [method, rref] + list(args)
+    return rpc.rpc_sync(rref.owner(), call_method, args=args, kwargs=kwargs)
 
 
-def run_trainer(ps_rref):
-    trainer = Trainer(ps_rref)
-    trainer.train()
+# --------- Parameter Server --------------------
+class ParameterServer(nn.Module):
+    def __init__(self, num_gpus=0):
+        super().__init__()
+        model = Net(num_gpus=num_gpus)
+        self.model = model
+        self.input_device = torch.device(
+            "cuda:0" if torch.cuda.is_available() and num_gpus > 0 else "cpu")
+
+    def forward(self, inp):
+        inp = inp.to(self.input_device)
+        out = self.model(inp)
+        # This output is forwarded over RPC, which as of 1.5.0 only accepts CPU tensors.
+        # Tensors must be moved in and out of GPU memory due to this.
+        out = out.to("cpu")
+        return out
+
+    # Use dist autograd to retrieve gradients accumulated for this model.
+    # Primarily used for verification.
+    def get_dist_gradients(self, cid):
+        grads = dist_autograd.get_gradients(cid)
+        # This output is forwarded over RPC, which as of 1.5.0 only accepts CPU tensors.
+        # Tensors must be moved in and out of GPU memory due to this.
+        cpu_grads = {}
+        for k, v in grads.items():
+            k_cpu, v_cpu = k.to("cpu"), v.to("cpu")
+            cpu_grads[k_cpu] = v_cpu
+        return cpu_grads
+
+    # Wrap local parameters in a RRef. Needed for building the
+    # DistributedOptimizer which optimizes paramters remotely.
+    def get_param_rrefs(self):
+        param_rrefs = [rpc.RRef(param) for param in self.model.parameters()]
+        return param_rrefs
 
 
-def run_ps(trainers):
-    timed_log("Start training")
-    start = perf_counter()
-    ps_rref = rpc.RRef(BatchUpdateParameterServer(len(trainers)))
-    futs = []
-    for trainer in trainers:
-        futs.append(
-            rpc.rpc_async(trainer, run_trainer, args=(ps_rref,))
-        )
+param_server = None
+global_lock = Lock()
 
-    torch.futures.wait_all(futs)
-    stop = perf_counter()
-    timed_log("Finish training")
-    timed_log(f"Time spent training: {stop-start}s")
+
+def get_parameter_server(num_gpus=0):
+    global param_server
+    # Ensure that we get only one handle to the ParameterServer.
+    with global_lock:
+        if not param_server:
+            # construct it once
+            param_server = ParameterServer(num_gpus=num_gpus)
+        return param_server
+
+
+def run_parameter_server(rank, world_size):
+    # The parameter server just acts as a host for the model and responds to
+    # requests from trainers, hence it does not need to run a loop.
+    # rpc.shutdown() will wait for all workers to complete by default, which
+    # in this case means that the parameter server will wait for all trainers
+    # to complete, and then exit.
+    print("PS master initializing RPC")
+    rpc.init_rpc(name="parameter_server", rank=rank, world_size=world_size)
+    print("RPC initialized! Running parameter server...")
+    rpc.shutdown()
+    print("RPC shutdown on parameter server.")
+
+
+# --------- Trainers --------------------
+
+# nn.Module corresponding to the network trained by this trainer. The
+# forward() method simply invokes the network on the given parameter
+# server.
+class TrainerNet(nn.Module):
+    def __init__(self, num_gpus=0):
+        super().__init__()
+        self.num_gpus = num_gpus
+        self.param_server_rref = rpc.remote(
+            "parameter_server", get_parameter_server, args=(num_gpus,))
+
+    def get_global_param_rrefs(self):
+        remote_params = remote_method(
+            ParameterServer.get_param_rrefs,
+            self.param_server_rref)
+        return remote_params
+
+    def forward(self, x):
+        model_output = remote_method(
+            ParameterServer.forward, self.param_server_rref, x)
+        return model_output
+
+
+def run_training_loop(rank, num_gpus, train_loader, test_loader):
+    # Runs the typical nueral network forward + backward + optimizer step, but
+    # in a distributed fashion.
+    net = TrainerNet(num_gpus=num_gpus)
+    # Build DistributedOptmizer.
+    param_rrefs = net.get_global_param_rrefs()
+    opt = DistributedOptimizer(optim.SGD, param_rrefs, lr=0.03)
+    for i, (data, target) in enumerate(train_loader):
+        with dist_autograd.context() as cid:
+            model_output = net(data)
+            target = target.to(model_output.device)
+            loss = F.nll_loss(model_output, target)
+            if i % 5 == 0:
+                print(f"Rank {rank} training batch {i} loss {loss.item()}")
+            dist_autograd.backward(cid, [loss])
+            # Ensure that dist autograd ran successfully and gradients were
+            # returned.
+            assert remote_method(
+                ParameterServer.get_dist_gradients,
+                net.param_server_rref,
+                cid) != {}
+            opt.step(cid)
+
+    print("Training complete!")
+    print("Getting accuracy....")
+    get_accuracy(test_loader, net)
+
+
+def get_accuracy(test_loader, model):
+    model.eval()
+    correct_sum = 0
+    # Use GPU to evaluate if possible
+    device = torch.device("cuda:0" if model.num_gpus > 0
+                                      and torch.cuda.is_available() else "cpu")
+    with torch.no_grad():
+        for i, (data, target) in enumerate(test_loader):
+            out = model(data)
+            pred = out.argmax(dim=1, keepdim=True)
+            pred, target = pred.to(device), target.to(device)
+            correct = pred.eq(target.view_as(pred)).sum().item()
+            correct_sum += correct
+
+    print(f"Accuracy {correct_sum / len(test_loader.dataset)}")
+
+
+# Main loop for trainers.
+def run_worker(rank, world_size, num_gpus, train_loader, test_loader):
+    # print(f"Worker rank {rank} initializing RPC")
+    # rpc.init_rpc(
+    #     name=f"trainer_{rank}",
+    #     rank=rank,
+    #     world_size=world_size)
+
+    # print(f"Worker {rank} done initializing RPC")
+
+    run_training_loop(rank, num_gpus, train_loader, test_loader)
+    # rpc.shutdown()
+
+
+class FakeMNIST(torch.utils.data.Dataset):
+    def __init__(self, len):
+        self.len = len
+        self.images = []
+        for i in range(10):
+            self.images.append(
+                torch.empty((1, 28, 28), dtype=torch.float).random_(0,
+                                                                    1000) / 1000)
+
+    def __len__(self):
+        return self.len
+
+    def __getitem__(self, idx):
+        target = idx % 10
+        img = self.images[target]
+        return img, target
+
 
 class ParameterServerTest(RpcAgentTestFixture):
+    @property
+    def world_size(self):
+        return 2
 
-    @dist_init(setup_rpc=False)
-    def test_batch_updating_parameter_server(self):
-
-        if self.rank != 0:
-            rpc.init_rpc(
-                name=worker_name(self.rank),
-                backend=self.rpc_backend,
-                rank=self.rank,
-                world_size=self.world_size,
-                rpc_backend_options=self.rpc_backend_options,
-            )
+    @property
+    def name(self):
+        if self.rank == 0:
+            return "parameter_server"
         else:
-            rpc.init_rpc(
-                name=worker_name(self.rank),
-                backend=self.rpc_backend,
-                rank=self.rank,
-                world_size=self.world_size,
-                rpc_backend_options=self.rpc_backend_options,
-            )
-            run_ps([f"{worker_name(r)}" for r in range(1, self.world_size)])
+            return super().name
 
-        rpc.shutdown()
+    @dist_init
+    def test_parameter_server(self):
+        if self.rank != 0:
+            # Get data to train on
+            mnist = FakeMNIST(3000)
+            train_loader = torch.utils.data.DataLoader(mnist, batch_size=32,
+                                                       shuffle=True)
+            test_loader = torch.utils.data.DataLoader(mnist, batch_size=32,
+                                                      shuffle=True)
+            run_worker(self.rank, self.world_size, 0, train_loader, test_loader)
